backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.110.0",
    "psycopg2-binary==2.9.9",
    "pydantic[email]==2.6.1",
    "pydantic-settings==2.2.1",
    "sqlalchemy==2.0.27",
    "uvicorn[standard]==0.27.1",
    "python-jose[cryptography]==3.3.0",
    "bcrypt>=4.0.0",
    "python-multipart==0.0.9",
    "alembic==1.13.1",
]
